requirements.txt
asgiref==3.8.1
boto3==1.37.12
botocore==1.37.12
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
django-email-users==2.1.0
django-extensions==3.2.3
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
geographiclib==2.0 
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.2
pillow==11.1.0
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
s3transfer==0.11.4
signals==0.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
